[mozart] Move display scaling of input events in Interpreter
Before being able to move all of InputReader, we need to move the
display scaling logic out of the launch instance.

Introducing a rudimentary API to register a display with the
InputInterpreter (For now, there is only one display)

Moving the display scaling of touch, mouse and stylus event inside their
respective state update methods.

Change-Id: I5781dc115484f7ae0572ae0295b2c748df22d19f