Implement new image buffer transfer mechanism.
This change eliminates page table thrashing and substantially improves
performance.

A Buffer consists of a vmo object, a fence eventpair, and a
retention eventpair.  The Buffer protocol specifies a mechanism for
pipelining, reusing buffers, and retaining mappings for active buffers.

BufferConsumer and BufferProducer manage the process of using
and recycling buffers in a fairly transparent way.  The buffers include
all of the necessary signalling in-band so they can be used as part of
any number of protocols.

Updated the libraries and examples to use this new mechanism.

In particular, Noodles now takes advantage of pipelining.

Change-Id: Id59795aadb7f5a2d1a04f413513b1fbc0787050c