Simplify the launcher.
The environment setup responsibilities have been moved to
//apps/modular/src/bootstrap.

The launcher now only deals with binding views to view trees.
If provided an application on the command-line, it will launch it
as before.  However if there are no arguments, the launcher will
keep running and simply offer the |mozart.Launcher| service to
its creator.

The launcher is designed to be more robust now in the event
that the compositor or view manager dies.

Change-Id: I3c5ddb9a8ce86a763c896549da9db894c50e97d4