[mozart] Fix fatal exception on exit
We had a race condition on exit in input, where the mouse would generate
a REMOVE event when the corresponding state was being destroyed. This
state would be destroyed when the root presenter was destroyed but was
still registered as a listener for events.

Refactor code to not generate events on destructor. Also switch input
interpreter listener to weak pointers.

Bug: MZRT-3
Change-Id: I614b2d8a8a656c772fba5a2b5a36e9dfa51f8c61