[mozart] Input Focus Part 2
Introduce the RequestFocus API on view containers.

The notion of input focus is owned by the view registry at this point
and input manager queries it to know where to dispatch keyboard events.

For now the input focus chain is the view branch on which the view sits,
but hoping to add APIs to manage the focus chain.

At this point the input manager still sets the focus on input down but we should move this in the future.

Change-Id: I182f18c061bf8e7105b6f2763bf060e816be4d94