[mozart] Fixing input lockup
When swiping from the edge of the device, first events would not find
any nodes in the scene graph and end up in a branch of the code that
would not pop the event out of the queue. When new events come in, it's
being processed only if there is just that one event, otherwise the
pipeline is supposed to take care of it.

Also fixing tests failing. Turns out that we run tests through
test_runner which needs to be made aware of the app quitting cleanly
otherwise it assumes that it died even though the tests passed.

Bug: MZRT-36
Change-Id: I54facf2279fa7df228621b7ed1e5e8ac4388a876