[mozart] Fix some issues with IME integration
Forgot to save the value of view_token when inserting in the map, so the
value was always 0 and we would never find the text input service for
any view.

There was also an issue in Reset in the text input service
implementation which did not check whether the bindings had been bound.

Added formatting too for TextSelection, TextRange and TextState to help
debug.

Also fixing
Bug: MZRT-43 #done

Change-Id: I4033c962938e5dfbf688eda7c75510fe53069401