Implement and use metrics events.
This change adds code to calculate the metrics of nodes in the scene
and deliver them to session listeners.

It also redefines view sizes and input pointer positions in terms
of logical pixel so that essentially all of Mozart is now defined
in the model coordinate system instead of the display coordinate
system.

This change can be landed once all Mozart clients are updated to use
the new definition.

Change-Id: Ieb42a4bc12894ff2b3b3b13319fe64a97379335e